Move local caching to runfo
All of the local caching of Azure data is now in runfo as this is the
only part of the stack that actually uses it
